+++ src/modules/Charts/hooks/useNewForm.js
@@ -1,0 +1,223 @@
+import {Form, notification} from "antd";
+import {useNavigate, useParams} from "react-router-dom";
+import {useDataEngine} from "@dhis2/app-runtime";
+import {useEffect, useState} from "react";
+import {useSelector} from "react-redux";
+import dayjs from "dayjs";
+import {findSectionObject} from "../helpers";
+import {useDataElements} from "./useGetDataElement";
+
+
+export const useNewForm = () => {
+
+    const [formSections, setFormSections] = useState({
+        patients: {},
+        antibiotics: {},
+        cultures: {},
+        dosage: {},
+        recommendation: {},
+        redFlags: {},
+        comments: {},
+        signature: {},
+    })
+
+    const [formValues, setFormValues] = useState([])
+
+    const [initialState, setInitialState] = useState({})
+
+    const [recommendationValues, setRecommendationValues] = useState([])
+
+    const [redFlagValues, setRedFlagValues] = useState([])
+
+    const [loading, setLoading] = useState(false)
+
+    const [chartDataLoading, setChartDataLoading] = useState(false)
+
+    const {stages, program} = useSelector(state => state.forms)
+
+    const {id: orgUnitID} = useSelector(state => state.orgUnit)
+
+    const {getDataElementByID} = useDataElements()
+
+    const [form] = Form.useForm()
+
+    const {eventId} = useParams()
+
+    const engine = useDataEngine()
+
+    const navigate = useNavigate()
+
+    const getChart = async () => {
+        try {
+            setChartDataLoading(true)
+
+            const response = await engine.query({
+                events: {
+                    resource: `tracker/events/${eventId}`
+                }
+            })
+
+            const dataValues = response.events.dataValues
+
+            dataValues.forEach(dataValue => {
+                const newObject = {}
+
+                if ((new Date(dataValue.value) == "Invalid Date") && isNaN(new Date(dataValue.value)))
+                    newObject[dataValue.dataElement] = dataValue.value
+                else
+                    newObject[dataValue.dataElement] = dayjs(dataValue.value)
+
+                setInitialState(prevState => ({
+                    ...prevState, ...newObject
+                }))
+            })
+
+
+        } catch (e) {
+            notification.error({
+                message: "error",
+                description: "Something went wrong"
+            })
+        } finally {
+            setChartDataLoading(false)
+        }
+    }
+
+    const onFinish = async values => {
+        const dataValues = Object.keys(values).map(key => ({
+            dataElement: key,
+            value: values[key]
+        }))
+
+        recommendationValues.forEach(value => {
+            dataValues.push({
+                dataElement: value,
+                value: true,
+            })
+        })
+
+        redFlagValues.forEach(value => {
+            dataValues.push({
+                dataElement: value,
+                value: true,
+            })
+        })
+
+        const payload = {
+            events: [
+                {
+                    "occurredAt": new Date().toJSON().slice(0, 10),
+                    "notes": [],
+                    program,
+                    "programStage": stages[0].id,
+                    orgUnit: orgUnitID,
+                    event: eventId,
+                    dataValues
+                }
+            ]
+        }
+
+        try {
+            setLoading(true)
+
+            const response = await engine.mutate({
+                resource: "tracker",
+                type: "create",
+                data: payload,
+                params: {
+                    async: false,
+                    importStrategy: "UPDATE",
+                    partial: true
+                }
+            })
+
+            if (response?.status === "OK")
+                navigate("/charts")
+
+        } catch (e) {
+            notification.error({
+                message: "error",
+                description: "Something went wrong"
+            })
+        } finally {
+            setLoading(false)
+        }
+
+    }
+
+    const onFieldsChange = (changedFields, allFields) => {
+        setFormValues(allFields.map(field => ({name: field.name[0], value: field.value})))
+    }
+
+    const checkIfValid = dataElementID => {
+        const formValues = form.getFieldsValue()
+
+        if (formValues === {})
+            return {validity: true}
+
+        const dataElementObject = getDataElementByID(dataElementID)
+
+        if (dataElementObject.attributeValues.length === 0)
+            return {validity: true}
+
+        let validity = true
+
+        dataElementObject.attributeValues.forEach(attribute => {
+            const attributeElementId = attribute.value.split(",")[0]
+
+            const attributeValue = attribute.value.split(",")[1]
+
+            const conditionalCheck = form.getFieldValue(attributeElementId)
+
+            if (conditionalCheck === undefined)
+                validity = validity
+            else
+                validity = validity && (conditionalCheck == attributeValue)
+        })
+
+        if (validity === false)
+            form.setFieldValue(dataElementID, "N/A")
+
+        return {validity}
+
+    }
+
+
+    useEffect(() => {
+        getChart()
+    }, [eventId]);
+
+
+    useEffect(() => {
+        if (stages?.length > 0)
+            setFormSections({
+                patients: findSectionObject({searchString: "Patients", sectionArray: stages[0].sections}),
+                antibiotics: findSectionObject({searchString: "Antibiotics", sectionArray: stages[0].sections}),
+                cultures: findSectionObject({searchString: "Cultures", sectionArray: stages[0].sections}),
+                dosage: findSectionObject({searchString: "Dosage", sectionArray: stages[0].sections}),
+                recommendation: findSectionObject({searchString: "Recommendation", sectionArray: stages[0].sections}),
+                redFlags: findSectionObject({searchString: "Flags", sectionArray: stages[0].sections}),
+                comments: findSectionObject({searchString: "Comments", sectionArray: stages[0].sections}),
+                signature: findSectionObject({searchString: "Signature", sectionArray: stages[0].sections}),
+            })
+    }, [stages]);
+
+    return {
+        formSections,
+        formValues,
+        setFormValues,
+        initialState,
+        setRecommendationValues,
+        redFlagValues,
+        setRedFlagValues,
+        loading,
+        chartDataLoading,
+        form,
+        eventId,
+        getChart,
+        onFinish,
+        checkIfValid,
+        onFieldsChange
+    }
+
+}
